add: 拆分状态栏更新任务，使用任务通知或条件更新，减少LVGL API调用频率

+++ main/ui/status_bar.h
@@ -4,4 +4,4 @@
 void status_bar_create(void);
 void status_bar_set_time(const char *time_str);
 void status_bar_set_wifi_state(bool connected);
-void status_bar_set_battery_level(int level);
+void status_bar_set_battery_level(float level);
